Update validation API

+++ iOSFramework/iOSRuleEngine/iOSRuleEngine/Bridge.h
@@ -39,7 +39,7 @@
 -(void)setDoubleValue:(NSNumber*)value forPE:(unsigned short) peOid;
 -(void)setBoolValue:(NSNumber*)value forPE:(unsigned short) peOid;
 
--(void)addContributionStep:(int8_t)index employeePct:(double)employeePct companyPct:(double)companyPct;
+-(void)addContributionStep:(long)index employeePct:(double)employeePct companyPct:(double)companyPct;
 -(void)removeAllContributionSteps;
 
 -(void)unsetPEOid:(unsigned short) peOid;
